This is for python seving of the image after receiving it
Solved by turning the numpy image into a base64 by first converting it to PIL image , then writing it to a Bytestream object and then converting the bystream to a base64. Mark that this changes the values and the base64 of the original image but the resultant is same when look at the actual image.